Move SLF4J 1.x provider to its own module
