+++ .github/workflows/ci.yml
@@ -76,11 +76,11 @@ jobs:
 
       - name: Make target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
-        run: mkdir -p core/target zio/target extras/target json/target slf4j/target project/target
+        run: mkdir -p core/target zio/target extras/target json/target slf4j/target slf4j1-provider/target project/target
 
       - name: Compress target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
-        run: tar cf targets.tar core/target zio/target extras/target json/target slf4j/target project/target
+        run: tar cf targets.tar core/target zio/target extras/target json/target slf4j/target slf4j1-provider/target project/target
 
       - name: Upload target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
