+++ slf4j1-provider/src/main/scala-2/io/odin/slf4j/OdinLoggerVarargsAdapter.scala
