+++ slf4j1-provider/src/main/scala-3/io/odin/slf4j/OdinLoggerVarargsAdapter.scala
