+++ slf4j1-provider/src/main/scala/io/odin/slf4j/OdinLoggerAdapter.scala
