+++ slf4j1-provider/src/test/java/org/slf4j/impl/StaticLoggerBinder.java
