+++ slf4j1-provider/src/test/scala/io/odin/slf4j/ExternalLogger.scala
@@ -16,8 +16,6 @@
 
 package io.odin.slf4j
 
-import scala.annotation.nowarn
-
 import io.odin.{Level, Logger}
 
 import cats.effect.kernel.Sync
@@ -25,7 +23,6 @@ import cats.effect.std.Dispatcher
 import cats.effect.unsafe.implicits.global
 import cats.effect.IO
 
-@nowarn("cat=deprecation&msg=This logger will be moved to internal library test sources")
 class ExternalLogger extends OdinLoggerBinder[IO] {
 
   implicit val F: Sync[IO]                = IO.asyncForIO
